Make the command prompt a little more thread-safe
This one was accomplished by spliting the internal and external
control codes for signaling beggining and end of user input. The
command prompt gets drawn only when the external control code is
received, not when the listeners get done processing (they might
have spun other threads that still want to output stuff to the
console).